amélioration après correction rentmyride2

+++ controllers/dashboard/vehicles/deleteVehicles-ctrl.php
@@ -4,19 +4,22 @@ require_once(__DIR__ . '/../../../models/Vehicle.php');
 
 try {
 
-    // Récupération de l'ID du véhicule et nettoyage
+    // Récupération du paramètre d'URL correspondant à l'id de la catégorie cliquée
     $idVehicle = intval(filter_input(INPUT_GET, 'id_vehicle', FILTER_SANITIZE_NUMBER_INT));
 
     // Récupération des informations du véhicule
     $vehicle = Vehicle::get($idVehicle);
 
-    // suppression du véhicule
-    $isDeleted = Vehicle::delete($idVehicle);
+    if ($vehicle) {
 
-    if ($isDeleted) {
+        // Appel de la méthode delete
+        $isDeleted = Vehicle::delete($idVehicle);
+        
+        // Suppression du visuel associé
         $path = __DIR__ . "/../../public/uploads/vehicles/$vehicle->picture";
         @unlink($path);
         $msg = 'Véhicule supprimé avec succès.';
+
     } else {
         $msg = 'Le véhicule n\'a pas été supprimé.';
     }
@@ -27,7 +30,7 @@ try {
 
     die;
 } catch (Throwable $e) {
-    $error = $th->getMessage();
+    $error = $e->getMessage();
     include __DIR__ . '/../../../views/dashboard/templates/header_dashboard.php';
     include __DIR__ . '/../../../views/dashboard/templates/error.php';
     include __DIR__ . '/../../../views/dashboard/templates/footer_dashboard.php';
